k256: factor apart benchmarks
This is a slightly better alternative to #368 in that we can gate
feature-dependent benchmarks on individual bench modules.

(Note: the larger goal here is to add ECDSA benchmarks... this is just
moving things around to make that easier)
